image_streaming_test/tools/image_streaming/requirements.txt
pyserial
Pillow
